+++ app/stores/WalletBackupStore.js
@@ -39,6 +39,7 @@ export default class WalletBackupStore extends Store {
   }
 
   @action _initiateWalletBackup = (params: { recoveryPhrase: Array<string> }) => {
+    console.debug('initiate backup');
     this.recoveryPhrase = params.recoveryPhrase;
     this.inProgress = true;
     this.currentStep = 'privacyWarning';