+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Changelog
 - Multiple input and output addresses in transaction details
 - Ada redemption acceptance tests
 - Show BTC and ETC currencies as coming soon in create wallet dialog
+- Exporting paper wallets
 
 ### Fixes